Sectore_2 (2010)
The 2nd map in the SectorE series. Does not have any bug fixes or additions made by the Sven Co-op Team.

This map was created for an older version of Sven Co-op, but still is probably not fully compatable with original GoldSource in it's current state. Map changes and events/triggers might be buggy or not work.