Sectore_3 (2010)
The 3rd map concluding the first SectorE series. Does not have any bug fixes or additions made by the Sven Co-op Team.

This map was created for an older version of Sven Co-op, but still is probably not fully compatable with original GoldSource in it's current state. Map changes and events/triggers might be buggy or not work.

*** KNOWN BUGS ***
The map may crash if you jump from the ladder in the shaft to the elevator. func_ladders+half-life1code=funky things. hopefully solokiller can take a look at revamping the func_ladder code when he gets around to it for his Half-Life EnhancedSDK! :D